+++ default.nix
@@ -2,7 +2,7 @@ let
   # To upgrade rib, go to https://github.com/srid/rib/commits/master, select the
   # revision you would like to upgrade to and set it here. Consult rib's
   # ChangeLog.md to check any notes on API migration.
-  ribRevision = "77745c0";
+  ribRevision = "c115fe3";
   projectRoot = ./.;
 in {
 # Rib library source to use
